requirements.txt
firebase-admin
python-telegram-bot==13.7
